Files not directly attached to Model-to-be-Printed
This contains edited files not directly placed as an editable file in the Google Classroom Model to be Printed assignment.  Files include some formatting corrections, as well as my comments ("/**/") on what I would have liked to have seen, what was confusing, etc.  If you received something about seeing comments on GitHub in your "Returned" assignment, this is where you should be looking.